Moved function to add the MPRO amino acid sequence to PDB header from add_seqres.py to datasets.utils. Moved add_seqres.py script to scripts directory.
